requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appopener==1.7
certifi==2024.12.14
distro==1.9.0
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
MouseInfo==0.1.3
ollama==0.4.7
psutil==6.1.1
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
pytweening==1.2.0
pywin32==308
regex==2024.11.6
sniffio==1.3.1
typing_extensions==4.12.2
wheel==0.45.1
